feat : location get insert in homecontroller

+++ lib/app/modules/home/infrastructure/models/location_find_model.dart
@@ -1,5 +1,5 @@
 import 'package:agence_teste/app/modules/home/domain/entities/location_result.dart';
 
 class LocationFind extends LocationResult {
-  LocationFind(int lat, int long) : super(lat, long);
+  LocationFind(double lat, double long) : super(lat, long);
 }
